User Service open-feign: (4) Review 2

+++ src/main/java/ru/practicum/workshop/userservice/service/UserServiceImpl.java
@@ -1,26 +1,22 @@
 package ru.practicum.workshop.userservice.service;
 
-import feign.FeignException;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import ru.practicum.workshop.userservice.dto.AuthRegistrationDto;
 import ru.practicum.workshop.userservice.dto.UpdateUserFromRegistrationDto;
 import ru.practicum.workshop.userservice.dto.NewUserDto;
-import ru.practicum.workshop.userservice.dto.ResponseWithUserId;
 import ru.practicum.workshop.userservice.dto.UpdateUserDto;
 import ru.practicum.workshop.userservice.dto.UserDto;
-import ru.practicum.workshop.userservice.enums.RegistrationType;
+import ru.practicum.workshop.userservice.model.enums.RegistrationType;
 import ru.practicum.workshop.userservice.exception.AuthenticationException;
 import ru.practicum.workshop.userservice.mapping.UserMapper;
 import ru.practicum.workshop.userservice.model.User;
 import ru.practicum.workshop.userservice.repository.UserRepository;
 
 import java.util.List;
-import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -29,11 +25,19 @@ import java.util.Optional;
 public class UserServiceImpl implements UserService {
     private final UserRepository userRepository;
     private final UserMapper userMapper;
-    private final RegistrationClient registrationClient;
 
     @Override
     @Transactional
     public UserDto createUser(NewUserDto newUserDto) {
+        Optional<User> availableUser = userRepository.getUserByEmail(newUserDto.getEmail());
+        if (availableUser.isPresent() && availableUser.get().getRegistrationType().equals(RegistrationType.AUTO)) {
+            User userForTransferToManual = availableUser.get();
+            userForTransferToManual.setRegistrationType(RegistrationType.MANUAL);
+            User userTransferredToManual = userRepository.save(userMapper
+                    .changeFieldsOfAutoUserToManual(userForTransferToManual, newUserDto));
+            return userMapper.toUserDtoPrivate(userTransferredToManual);
+        }
+
         User newUser = userRepository.save(userMapper.toUser(newUserDto, RegistrationType.MANUAL));
 
         log.info("User added: {}", newUser);
@@ -43,19 +47,14 @@ public class UserServiceImpl implements UserService {
 
     @Override
     @Transactional
-    public ResponseWithUserId autoCreateUser(NewUserDto newUserDto) {
-
-        User availableUser = userRepository.getUserByEmail(newUserDto.getEmail());
-
-        if (availableUser != null && availableUser.getId() > 0) {
-            return new ResponseWithUserId(availableUser.getId());
+    public Long autoCreateUser(NewUserDto newUserDto) {
+        Optional<User> availableUser = userRepository.getUserByEmail(newUserDto.getEmail());
+        if (availableUser.isPresent()) {
+            return availableUser.get().getId();
         }
-
         User newUser = userRepository.save(userMapper.toUser(newUserDto, RegistrationType.AUTO));
-
         log.info("Auto user added: {}", newUser);
-
-        return new ResponseWithUserId(newUser.getId());
+        return newUser.getId();
     }
 
     @Override
@@ -78,40 +77,11 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional
     public void autoUpdateUserData(UpdateUserFromRegistrationDto updateUserFromRegistrationDto, Long userId) {
-        User userToUpdate = userRepository.getReferenceById(userId);
-
-        if (userToUpdate.getRegistrationType().equals(RegistrationType.MANUAL)) {
-            throw new AuthenticationException("User with id=" + userId + " have MANUAL registration, and can't be" +
-                    " updated from registration-service.");
+        Optional<User> userToUpdate = userRepository.findById(userId);
+        if (userToUpdate.isPresent() && userToUpdate.get().getRegistrationType().equals(RegistrationType.AUTO)) {
+            User updatedUser = userRepository.save(userMapper.autoUpdateUser(userToUpdate.get(), updateUserFromRegistrationDto));
+            log.info("Auto user updated: {}", updatedUser);
         }
-
-        User updatedUser = userRepository.save(userMapper.autoUpdateUser(userToUpdate, updateUserFromRegistrationDto));
-
-        log.info("Auto user updated: {}", updatedUser);
-
-    }
-
-    @Override
-    public void transferUserToManual(Long userId, AuthRegistrationDto authRegistrationDto) {
-        User userForTransfer = userRepository.findById(userId).orElseThrow(
-                () -> new AuthenticationException("Service don't have information about user with id=" + userId));
-        if (userForTransfer.getRegistrationType().equals(RegistrationType.MANUAL)) {
-            throw new AuthenticationException("Registration status of user with id=" + userId + " is already MANUAL.");
-        }
-
-        ResponseWithUserId responseWithUserId;
-        try {
-            responseWithUserId = registrationClient.confirmUser(authRegistrationDto.getId(), authRegistrationDto.getPassword());
-        } catch (FeignException.NotFound e) {
-            throw new AuthenticationException("Registration service don't have information about registration with id=" + authRegistrationDto.getId());
-        }
-
-        if (!(Objects.equals(responseWithUserId.getUserId(), userId))) {
-            throw new AuthenticationException("Registration service don't have information about registration with id=" + authRegistrationDto.getId() +
-                    " from user with id=" + userId);
-        }
-        userForTransfer.setRegistrationType(RegistrationType.MANUAL);
-        userRepository.save(userForTransfer);
     }
 
     @Override
@@ -132,8 +102,8 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional
     public void autoDeleteUser(Long userId) {
-        User userForAutoDelete = userRepository.getReferenceById(userId);
-        if (userForAutoDelete.getRegistrationType().equals(RegistrationType.AUTO)) {
+        Optional<User> userForAutoDelete = userRepository.findById(userId);
+        if (userForAutoDelete.isPresent() && userForAutoDelete.get().getRegistrationType().equals(RegistrationType.AUTO)) {
             log.info("Auto user deleted: id={}", userId);
             userRepository.deleteById(userId);
         }
